rename module lazy to lazylock
