+++ src/lazylock.rs
